Improve mobile nav and add event selector to nav
