+++ package.json
@@ -52,7 +52,7 @@
         "react-redux": "^8.1.3",
         "use-firebase-auth": "^3.0.0",
         "uuid": "^9.0.1",
-        "wouter": "^2.12.1",
+        "wouter": "^3.3.5",
         "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.2/xlsx-0.20.2.tgz",
         "xlsx-js-style": "^1.2.0",
         "yup": "^1.3.2"
