+++ src/events/page/EventRouter.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { lazy, Suspense, useEffect } from 'react'
 import { Redirect, Route, Switch, useRoute } from 'wouter'
 import { NestedRoutes } from '../../components/NestedRoutes'
-import { EventLayout } from './EventLayout'
+import { EventLayout } from './layouts/EventLayout'
 import { useEvent } from '../../services/hooks/useEvent'
 import { FirestoreQueryLoaderAndErrorDisplay } from '../../components/FirestoreQueryLoaderAndErrorDisplay'
 import { Event } from '../../types'
